mail-summarizer/requirements.txt
kafka-python
langchain
langchain-community
